refactor: Roll all OpenAI-related functions into OpenAICompletionService
