+++ packages/navie/src/lib/oracle.ts
@@ -1,26 +1,18 @@
-import { ChatOpenAI } from '@langchain/openai';
 import { warn } from 'console';
-import OpenAI from 'openai';
 
 import Message from '../message';
-import completion from './completion';
+import CompletionService from '../services/completion-service';
 
 export default class Oracle {
   constructor(
     public name: string,
     private systemPrompt: string,
-    private modelName: string,
-    private temperature: number
+    private completionService: CompletionService
   ) {}
 
   async ask(messages: Message[], question?: string): Promise<string | undefined> {
-    const openAI: ChatOpenAI = new ChatOpenAI({
-      modelName: this.modelName,
-      temperature: this.temperature,
-    });
-
-    const collectMessages = (): OpenAI.ChatCompletionMessageParam[] => {
-      const result: OpenAI.ChatCompletionMessageParam[] = [
+    const collectMessages = (): Message[] => {
+      const result: Message[] = [
         {
           content: this.systemPrompt,
           role: 'system',
@@ -37,7 +29,7 @@ export default class Oracle {
     };
 
     let response = '';
-    for await (const token of completion(openAI, collectMessages())) {
+    for await (const token of this.completionService.complete(collectMessages())) {
       response += token;
     }
